TestUtils/UtilityMethodTestCase: split a test class
Split the `UtilityMethodTestCaseTest` in two separate test classes, one for the setup, one for the teardown.
